requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
httpx~=0.28
redis>=5.0
pydantic-settings>=2.2
prometheus-client>=0.20
opentelemetry-api~=1.36
opentelemetry-sdk~=1.36
opentelemetry-instrumentation-asgi~=0.57b0
opentelemetry-instrumentation-sqlalchemy~=0.57b0
opentelemetry-instrumentation-fastapi~=0.57b0
opentelemetry-exporter-otlp-proto-http~=1.36
